Move `CreateSelectColumnsForInstances` to a separate file.
